requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
httpx
azure-storage-blob
azure-cosmos
azure-cognitiveservices-speech
azure-ai-vision-imageanalysis
azure-ai-contentsafety
pydantic
joblib
openai
pillow
pymupdf
youtube-transcript-api
assemblyai
httpx
yt-dlp
numpy
xgboost
scikit-learn